requirements.txt
ipython
google-cloud-vision
